org-ubid relationship redundant
