+++ seed/models/ubids.py
@@ -11,4 +11,4 @@ class Ubid(models.Model):
     property = models.ForeignKey(PropertyState, on_delete=models.CASCADE, null=True)
     taxlot = models.ForeignKey(TaxLotState, on_delete=models.CASCADE, null=True)
     preferred = models.BooleanField(default=False)
-    organization = models.ForeignKey(Organization, on_delete=models.CASCADE, blank=False)
+
